requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
alembic==1.13.1
flask-jwt-extended
flask-bcrypt